+++ src/default_package_config.rs
@@ -1,5 +1,4 @@
 
-
 pub static DEFAULT_PACKAGE_CONFIG: &str = r#"
 # This is a configuration file for the bacon tool
 # More info at https://github.com/Canop/bacon
@@ -9,6 +8,10 @@ default_job = "check"
 [jobs]
 
 [jobs.check]
+command = ["cargo", "check", "--tests", "--color", "always"]
+need_stdout = false
+
+[jobs.light]
 command = ["cargo", "check", "--color", "always"]
 need_stdout = false